// benefits-frontend/src/api/schemas/offerStatus.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Employee Benefits API
 * OpenAPI spec version: 0.1.0
 */

export type OfferStatus = (typeof OfferStatus)[keyof typeof OfferStatus];

export const OfferStatus = {
  DRAFT: "DRAFT",
  ACTIVE: "ACTIVE",
  EXPIRED: "EXPIRED",
} as const;