// benefits-frontend/src/api/schemas/partnerOfferListItemResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Employee Benefits API
 * OpenAPI spec version: 0.1.0
 */
import type { OfferStatus } from "./offerStatus";

export interface PartnerOfferListItemResponse {
  id: string;
  title: string;
  status: OfferStatus;
  validFrom: string;
  validTo: string;
  /** @minimum 0 */
  availableCodes: number;
  /** @minimum 0 */
  totalCodes: number;
}